readelf: Solve comparison between signed and unsigned compiler warning.

+++ src/readelf.c
@@ -4638,7 +4638,7 @@ print_debug_ranges_section (Dwfl_Module *dwflmod,
 				      offset, &readp, endp))
 	continue;
 
-      if (unlikely (data->d_size - offset < address_size * 2))
+      if (unlikely (data->d_size - offset < (size_t) address_size * 2))
 	{
 	  printf (gettext (" [%6tx]  <INVALID DATA>\n"), offset);
 	  break;
@@ -6554,7 +6554,7 @@ print_debug_loc_section (Dwfl_Module *dwflmod,
 				      &cu, offset, &readp, endp))
 	continue;
 
-      if (unlikely (data->d_size - offset < address_size * 2))
+      if (unlikely (data->d_size - offset < (size_t) address_size * 2))
 	{
 	  printf (gettext (" [%6tx]  <INVALID DATA>\n"), offset);
 	  break;
